day03/requirements_dna.txt
biopython
pandas
pytest
numpy
